app/requirements.txt
Flask==3.0.0
psutil>=7.1.1
mysql-connector-python>=9.0.0